Rename component to be more explicit
