+++ react-app/src/components/TestTemplate/SectionAndQuestionTypes.js
